Added a lot of comments
Fixed the initial learning rate for ADADELTA and added some diagnostics, as well as a call operator for making predictions.

PREVIOUSLY AND STILL RELEVANT
Added dropout - just need file read/write for the basic FFNetwork and then it's time to move on to CUDA implementations.

Also working on commenting.

Finally - added the mnist dataset and a reader from Michael Nielsen's Neural Networks and Deep Learning book/post thing.  Will be updating it to feed into my network design but for now it's all good.